fix: allow casting `u128` in a comptime context
